+++ src/Task/Changelog.php
@@ -1,6 +1,8 @@
 <?php
 namespace Robo\Task;
 
+use Robo\Result;
+
 trait Changelog
 {
     protected function taskChangelog($filename = 'CHANGELOG.md')
@@ -9,10 +11,18 @@ trait Changelog
     }
 }
 
+/**
+ * @method ChangelogTask filename(string $filename)
+ * @method ChangelogTask anchor(string $anchor)
+ * @method ChangelogTask version(string $version)
+ *
+ * @package Robo\Task
+ */
 class ChangelogTask implements TaskInterface
 {
     use \Robo\Output;
     use FileSystem;
+    use DynamicConfig;
 
     protected $filename;
     protected $log = [];
@@ -21,7 +31,7 @@ class ChangelogTask implements TaskInterface
 
     public function askForChanges()
     {
-        while ($resp = $this->ask("Changed in this release:")) {
+        while ($resp = $this->ask("Changed in this release: ")) {
             $this->log[] = $resp;
         };
         return $this;
@@ -37,22 +47,16 @@ class ChangelogTask implements TaskInterface
         $this->log = array_merge($this->log, $data);
         return $this;
     }
-    
-    public function anchor($anchor)
+
+    public function change($change)
     {
-        $this->anchor = $anchor;
+        $this->log[] = $change;
         return $this;
     }
-
+    
     public function getChanges()
     {
         return $this->log;
-    }
-
-    public function version($version)
-    {
-        $this->version = $version;
-        return $this;
     }
 
     public function run()
@@ -61,18 +65,30 @@ class ChangelogTask implements TaskInterface
             $this->printTaskInfo("<alert>Changelog is empty</alert>");
             return false;
         }
-        $text = implode("\n", array_map(function ($i) { return "* $i"; }, $this->log));
-        $text = "#### {$this->version} ".date('m/d/Y')."\n\n".$text;
+        $text = implode("\n", array_map(function ($i) { return "* $i"; }, $this->log))."\n";
+        $ver = "#### {$this->version} ".date('m/d/Y')."\n\n";
+        $text = $ver . $text;
 
         if (!file_exists($this->filename)) {
             $this->printTaskInfo("Creating {$this->filename}");
-            file_put_contents($this->filename, $this->anchor);
+            $res = file_put_contents($this->filename, $this->anchor);
+            if ($res === false) return Result::error($this, "File {$this->filename} cant be created");
         }
 
-        return (new ReplaceInFileTask($this->filename))
-            ->from($this->anchor)
-            ->to($this->anchor."\n\n".$text)
+        // trying to append to changelog for today
+        $result = (new ReplaceInFileTask($this->filename))
+            ->from($ver)
+            ->to($text)
             ->run();
+
+        if (!$result->getData()['replaced']) {
+            $result = (new ReplaceInFileTask($this->filename))
+                ->from($this->anchor)
+                ->to($this->anchor."\n\n".$text)
+                ->run();
+        }
+
+        return new Result($this, $result->getExitCode(), $result->getMessage(), $this->log);
 
     }