+++ src/Task/GitHub.php
@@ -1,6 +1,7 @@
 <?php
 namespace Robo\Task;
 
+use Robo\Result;
 use Symfony\Component\Console\Helper\DialogHelper;
 const GITHUB_URL = 'https://api.github.com';
 
@@ -16,9 +17,14 @@ trait GitHub
     }
 }
 
+/**
+ * @method GitHubTask repo(string)
+ * @method GitHubTask owner(string)
+ */
 abstract class GitHubTask implements TaskInterface
 {
     use \Robo\Output;
+    use DynamicConfig;
 
     protected static $user;
     protected static $pass;
@@ -26,18 +32,6 @@ abstract class GitHubTask implements TaskInterface
     protected $needs_auth = false;
     protected $repo;
     protected $owner;
-
-    public function repo($repo)
-    {
-        $this->repo = $repo;
-        return $this;
-    }
-
-    public function owner($owner)
-    {
-        $this->owner = $owner;
-        return $this;
-    }
 
     public function uri($uri)
     {
@@ -75,21 +69,33 @@ abstract class GitHubTask implements TaskInterface
             CURLOPT_FOLLOWLOCATION => true,
             CURLOPT_USERAGENT => self::$user ?: "Robo"
         ));
-        
+
+        $code = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         $output = curl_exec($ch);
         $response = json_decode($output);
 
-        if ($response->message == "Not Found" and !self::$user) {
+        if ($code == 404 and !self::$user) {
             $this->askAuth();
             curl_setopt($ch, CURLOPT_USERPWD, self::$user.':'.self::$pass);
             $output = curl_exec($ch);
             $response = json_decode($output);
         }
         $this->printTaskInfo($output);
-        return $response;
+        return [$code, $response];
     }
 }
 
+/**
+ * @method GitHubReleaseTask tag(string)
+ * @method GitHubReleaseTask name(string)
+ * @method GitHubReleaseTask body(string)
+ * @method GitHubReleaseTask draft(boolean)
+ * @method GitHubReleaseTask prerelease(boolean)
+ * @method GitHubReleaseTask comittish(string)
+ *
+ * Class GitHubReleaseTask
+ * @package Robo\Task
+ */
 class GitHubReleaseTask extends GitHubTask implements TaskInterface
 {
     protected $tag;
@@ -102,12 +108,6 @@ class GitHubReleaseTask extends GitHubTask implements TaskInterface
     public function __construct($tag)
     {
         $this->tag = $tag;
-    }
-
-    public function desc($body)
-    {
-        $this->body = $body;
-        return $this;
     }
 
     public function askName()
@@ -140,39 +140,9 @@ class GitHubReleaseTask extends GitHubTask implements TaskInterface
         return $this;
     }
 
-    public function draft($draft = true)
-    {
-        $this->draft = $draft;
-        return $this;
-    }
-
-    public function name($name)
-    {
-        $this->name = $name;
-        return $this;
-    }
-
-    public function prerelease($prerelease = true)
-    {
-        $this->prerelease = $prerelease;
-        return $this;
-    }
-
-    public function tag($tag)
-    {
-        $this->tag = $tag;
-        return $this;
-    }
-
-    public function committish($branchOrSha)
-    {
-        $this->comittish = $branchOrSha;
-        return $this;
-    }
-
     public function run()
     {
-        $this->sendRequest('releases', [
+        list($code, $data) = $this->sendRequest('releases', [
             "tag_name" => $this->tag,
             "target_commitish" => $this->comittish,
             "name" => $this->tag,
@@ -180,5 +150,12 @@ class GitHubReleaseTask extends GitHubTask implements TaskInterface
             "draft" => $this->draft,
             "prerelease" => $this->prerelease
         ]);
+
+        return new Result(
+            $this,
+            in_array($code, [200, 201]) ? 0 : 1,
+            isset($data->message) ? $data->message : '',
+            $data
+        );
     }
 }