+++ src/Task/TaskInterface.php
@@ -1,6 +1,12 @@
 <?php
 namespace Robo\Task;
 
+use Robo\Result;
+
 interface TaskInterface {
+
+    /**
+     * @return Result
+     */
     function run();
 }