+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+#### 0.3.0 02/11/2014
+
+* Dynamic configuration via magic methods
+* added WriteToFile task
+* Result class for managing exit codes and error messages
+
 #### 0.2.0 01/29/2014
 
 * Merged Tasks and Traits to same file